Added Fenwick2D tree for range updates

+++ src/com/br/uri/christmas/contest2017/ChristmasLights.java
@@ -5,6 +5,7 @@ import java.util.Scanner;
 /**
  * Created by rene on 16/12/17.
  */
+//https://www.urionlinejudge.com.br/judge/en/problems/view/2718
 //https://www.urionlinejudge.com.br/judge/en/challenges/view/338/2
 public class ChristmasLights {
 
@@ -13,7 +14,7 @@ public class ChristmasLights {
         int groups = scanner.nextInt();
 
         for(int group = 0; group < groups; group++) {
-            long bulbs = scanner.nextInt();
+            long bulbs = scanner.nextLong();
 
             String binaryBulbs = "";
